Добавлены файлы:QuizStepViewModel.swift, QuizResultsViewModel.swift, QuizQuestion.swift
