--- MovieQuiz/QuizStepViewModel.swift
@@ -1,8 +1,0 @@
-//
-//  QuizStepViewModel.swift
-//  MovieQuiz
-//
-//  Created by Алина on 30.11.2024.
-//
-
-import Foundation
